feat(isCgi): move isCgi to Client Class
