+++ src/Client.hpp
@@ -87,4 +87,5 @@ std::ostream& operator<<(std::ostream& out, Client& client);
 Client* findClient(int fd, std::vector<Client*>& clients);
 void	disconnectClient(int fd, std::vector<Client*>& clients,
 					  std::vector<pollfd>& fdPool, std::vector<pollfd>::iterator toRemove);
+bool	isCgi(std::string& uri, Location& location);
 #endif // !CLIENT_HPP
